Changes in get_database & find_matches
get_database was restructured to deal with potential errors when reinstalling the database. Also it was add a conditional to check if the function was called  directly or indirectly (from find_matches).

find_matches was updated to deal with the calling of get_database(); also, now best_matches_all.m8 and best_matches.m8 have a header row that works just fine.

In align_matches, the way best_matches.m8 was read was changed to use pandas (the code is now more concise and clear than before).

In the setup.py, I add "ete3" as library to load, as it was not in the list.